Weaver SyncVar processing moved into NetworkBehaviourSyncVarProcessor
